+++ worker.ts
@@ -1,0 +1,3 @@
+import { registerWorker as registerShakerWorker } from "./shaker/worker.ts";
+
+registerShakerWorker();